+++ tests/testthat.R
@@ -1,4 +1,0 @@
-library(testthat)
-library(nzffdr)
-
-test_check("nzffdr")