+++ tests/testthat/test-clean_nzffd.R
@@ -1,5 +1,0 @@
-test_that("nzffd_clean works", {
-  cd1 <- nzffd_clean(fishd = nzffd_data)
-  expect_equal(c(200,23), dim(cd1))
-})
-